refactor(react-icons-file-type): rename getFileTypeIconSrc module to fileTypeIconResolver
The module now owns the full icon-resolution pipeline (FileTypeIconOptions,
getFileTypeIconNameFromExtensionOrType, getFileTypeIconSuffix, getFileTypeIconSrc),
so its name no longer reflects its contents. Rename the source and test files;
function names are unchanged. Internal only — public API unaffected.
