+++ packages/react-icons-file-type/src/fileTypeIconResolver.ts
